adding shareClusterStatesAcrossEndpoints CLI flag to enable sync mode via CLI
BUG: ZAPP-1274
