+++ src-electron/rest/user-data.js
@@ -37,6 +37,7 @@ const restApi = require('../../src-shared/rest-api.js')
 const zclLoader = require('../zcl/zcl-loader.js')
 const dbEnum = require('../../src-shared/db-enum.js')
 const { StatusCodes } = require('http-status-codes')
+const clusterSync = require('../upgrade/cluster-sync')
 
 /**
  * HTTP GET: session key values
@@ -108,7 +109,8 @@ function httpGetPackageNotifications(db) {
 function httpGetPackageNotificationsByPackageId(db) {
   return async (request, response) => {
     let packageId = request.params.packageId
-    let notifications = await queryPackageNotification.getNotificationByPackageId(db, packageId)
+    let notifications =
+      await queryPackageNotification.getNotificationByPackageId(db, packageId)
     response.status(StatusCodes.OK).json(notifications)
   }
 }
@@ -136,7 +138,8 @@ function httpDeletePackageNotification(db) {
 function httpGetUnseenNotificationCount(db) {
   return async (request, response) => {
     let sessionId = request.zapSessionId
-    let notificationCount = await querySessionNotification.getUnseenNotificationCount(db, sessionId)
+    let notificationCount =
+      await querySessionNotification.getUnseenNotificationCount(db, sessionId)
     response.status(StatusCodes.OK).json(notificationCount)
   }
 }
@@ -553,262 +556,15 @@ function httpPostShareClusterStatesAcrossEndpoints(db) {
   return async (request, response) => {
     let { endpointTypeIdList } = request.body
     let sessionId = request.zapSessionId
-    let packageIds = await queryPackage.getSessionPackagesByType(
-      db,
-      sessionId,
-      dbEnum.packageType.zclProperties
-    )
-    packageIds = packageIds.map(function getId(item) {
-      return item.id
-    })
+
     if (!Array.isArray(endpointTypeIdList) || endpointTypeIdList.length < 1) {
       return response.status(StatusCodes.BAD_REQUEST).send()
     }
 
-    // Get a list of clusters enabled by multiple (>1) endpoints
-    let sharedClusterList = await queryEndpointType
-      .selectAllClustersDetailsFromEndpointTypes(
-        db,
-        endpointTypeIdList.map((id) => {
-          return { endpointTypeId: id }
-        })
-      )
-      .then((list) => list.filter((entry) => entry.endpointCount > 1))
+    let syncResult = clusterSync.clusterSync(db, sessionId, endpointTypeIdList)
 
-    let attrDefaults = await attributeDefaults(
-      db,
-      endpointTypeIdList,
-      sharedClusterList,
-      packageIds
-    )
-    await writeAttributeDefaults(db, attrDefaults)
-
-    let cmdDefaults = await commandDefaults(
-      db,
-      endpointTypeIdList,
-      sharedClusterList,
-      packageIds
-    )
-    await writeCommandDefaults(db, cmdDefaults)
-
-    return response.status(StatusCodes.OK).json({
-      sharedClusterList,
-      sharedAttributeDefaults: attrDefaults,
-    })
+    return response.status(StatusCodes.OK).json(syncResult)
   }
-}
-
-async function commandDefaults(
-  db,
-  endpointTypeIdList,
-  sharedClusterList,
-  packageIds
-) {
-  let sharedCmdDefaults = {}
-  let clusCmdToCmdObj = {}
-  let sharedCommandList =
-    await queryCommand.selectAllCommandDetailsFromEnabledClusters(
-      db,
-      sharedClusterList.map((c) => {
-        return { endpointTypeClusterRef: c.endpointClusterId }
-      }),
-      packageIds
-    )
-
-  for (const endpointTypeId of endpointTypeIdList) {
-    for (const sharedCmd of sharedCommandList) {
-      let clusCmdKey = JSON.stringify({
-        clusterId: sharedCmd.clusterId,
-        clusterSide: sharedCmd.clusterSide,
-        id: sharedCmd.id, // command id
-        code: sharedCmd.code,
-        mfgCode: sharedCmd.mfgCode,
-      })
-
-      if (!(endpointTypeId in sharedCmdDefaults)) {
-        sharedCmdDefaults[endpointTypeId] = []
-      }
-
-      if (clusCmdKey in clusCmdToCmdObj) {
-        sharedCmdDefaults[endpointTypeId].push(clusCmdToCmdObj[clusCmdKey])
-      } else {
-        let cmds = await queryEndpoint.selectEndpointClusterCommands(
-          db,
-          sharedCmd.clusterId,
-          endpointTypeId
-        )
-
-        // find attr
-        let matched = cmds.filter((cmd) => commandEquals(cmd, sharedCmd))
-        if (matched.length) {
-          let m = matched.shift()
-
-          sharedCmdDefaults[endpointTypeId].push(m)
-          clusCmdToCmdObj[clusCmdKey] = m
-        }
-      }
-    }
-  }
-  return sharedCmdDefaults
-}
-
-async function writeCommandDefaults(db, defaults) {
-  let promises = []
-  for (const [endpointTypeId, commandList] of Object.entries(defaults)) {
-    for (const cmd of commandList) {
-      promises.push(
-        queryConfig.insertOrUpdateCommandState(
-          db,
-          endpointTypeId,
-          cmd.clusterId,
-          cmd.source,
-          cmd.id,
-          cmd.isIncoming,
-          true
-        )
-      )
-
-      promises.push(
-        queryConfig.insertOrUpdateCommandState(
-          db,
-          endpointTypeId,
-          cmd.clusterId,
-          cmd.source,
-          cmd.id,
-          cmd.isOutgoing,
-          false
-        )
-      )
-    }
-  }
-  await Promise.all(promises)
-}
-
-async function attributeDefaults(
-  db,
-  endpointTypeIdList,
-  sharedClusterList,
-  packageIds
-) {
-  let sharedAttributeDefaults = {}
-  let clusterIdnSideToAttrCache = {}
-  let sharedAttributeList =
-    await queryAttribute.selectAttributeDetailsFromEnabledClusters(
-      db,
-      sharedClusterList,
-      packageIds
-    )
-
-  for (const endpointTypeId of endpointTypeIdList) {
-    for (const sharedAttr of sharedAttributeList) {
-      let clusAttrCacheKey = JSON.stringify({
-        clusterId: sharedAttr.clusterId,
-        side: sharedAttr.side,
-        id: sharedAttr.id, // attr id
-        code: sharedAttr.code,
-        name: sharedAttr.name,
-        type: sharedAttr.type,
-        mfgCode: sharedAttr.mfgCode,
-        define: sharedAttr.define,
-      })
-
-      if (clusAttrCacheKey in clusterIdnSideToAttrCache) {
-        !(endpointTypeId in sharedAttributeDefaults) &&
-          (sharedAttributeDefaults[endpointTypeId] = [])
-        sharedAttributeDefaults[endpointTypeId].push(
-          clusterIdnSideToAttrCache[clusAttrCacheKey]
-        )
-      } else {
-        let attributes = await queryEndpoint.selectEndpointClusterAttributes(
-          db,
-          sharedAttr.clusterId,
-          sharedAttr.side,
-          endpointTypeId
-        )
-
-        // find attr
-        let matched = attributes.filter((attr) =>
-          attributeEquals(attr, sharedAttr)
-        )
-        if (matched.length) {
-          let m = matched.shift()
-
-          !(endpointTypeId in sharedAttributeDefaults) &&
-            (sharedAttributeDefaults[endpointTypeId] = [])
-          sharedAttributeDefaults[endpointTypeId].push(m)
-          clusterIdnSideToAttrCache[clusAttrCacheKey] = m
-        }
-      }
-    }
-  }
-  return sharedAttributeDefaults
-}
-
-async function writeAttributeDefaults(db, defaults) {
-  let promises = []
-  for (const [endpointTypeId, attributeList] of Object.entries(defaults)) {
-    for (const attr of attributeList) {
-      promises.push(
-        queryConfig.insertOrUpdateAttributeState(
-          db,
-          endpointTypeId,
-          attr.clusterId,
-          attr.side,
-          attr.id,
-          [
-            { key: restApi.updateKey.attributeSelected, value: 1 },
-            {
-              key: restApi.updateKey.attributeStorage,
-              value: `"${attr.storage}"`,
-            },
-            {
-              key: restApi.updateKey.attributeSingleton,
-              value: attr.isSingleton,
-            },
-            {
-              key: restApi.updateKey.attributeBounded,
-              value: attr.isBounded,
-            },
-            {
-              key: restApi.updateKey.attributeDefault,
-              value: attr.defaultValue,
-            },
-            {
-              key: restApi.updateKey.attributeReporting,
-              value: attr.includedReportable,
-            },
-          ],
-          attr.min,
-          attr.max,
-          attr.reportableChange
-        )
-      )
-    }
-  }
-
-  await Promise.all(promises)
-}
-
-function commandEquals(a, b) {
-  return (
-    a.id === b.id &&
-    a.name === b.name &&
-    a.code === b.code &&
-    a.source === b.source &&
-    a.manufacturerCode === b.mfgCode
-  )
-}
-
-function attributeEquals(a, b) {
-  return (
-    a.id === b.id &&
-    a.code === b.code &&
-    a.name === b.name &&
-    a.side === b.side &&
-    a.type === b.type &&
-    a.manufacturerCode === b.mfgCode &&
-    a.define === b.define
-  )
 }
 
 function httpDeleteSessionPackage(db) {
